+++ lib/ex_insights/telemetry/metric_telemetry/metric.ex
@@ -15,6 +15,7 @@ defmodule ExInsights.Telemetry.MetricTelemetry.Metric do
           kind: kind()
         }
 
+  @derive Jason.Encoder
   @enforce_keys [:name, :value, :kind]
   defstruct [
     :name,